Remove unused common for now, Restructure Apollo Client into library
