+++ extensions/Cargo.toml
@@ -1,6 +1,4 @@
 [workspace]
-
 members = [
-    "common",
     "apollo_client",
 ]
