+++ extensions/apollo_client/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
-name = "apollo_client"
+name = "apollo_client_cli"
 version = "0.1.0"
 authors = ["TAC Mod Team <info@theseus-aegis.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "apollo_client"
+crate-type = ["cdylib"]
 
 [dependencies]
 arma-rs = { git = "https://github.com/synixebrett/arma-rs", rev = "4f0be98" }
-common = { path = "../common" }
 reqwest = { version = "0.10", features = ["blocking"] }
