--- extensions/common/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "common"
-version = "0.1.0"
-authors = ["TAC Mod Team <info@theseus-aegis.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
